PC-1679 fix show stock remaining
